#15 feat: 파티 생성 페이지 실내, 자연 암장 스텝 분리
